requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.1
pydantic-settings==2.0.3
email-validator==2.1.0.post1
fastapi-mail==1.4.1
aiosmtplib==2.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
